refactor: createHeader 메서드를 분리
